+++ lib/src/models/document_property_base.dart
@@ -30,7 +30,7 @@ library aspose_words_cloud;
 import '../../aspose_words_cloud.dart';
 
 /// Words document property DTO base class.
-class DocumentPropertyBase implements ModelBase {
+abstract class DocumentPropertyBase implements ModelBase {
   /// Gets or sets the value of the document property.
   String _value;