+++ lib/src/requests/accept_all_revisions_online_request.dart
@@ -89,7 +89,7 @@ class AcceptAllRevisionsOnlineRequest implements RequestBase {
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
     var _result = AcceptAllRevisionsOnlineResponse();
-    _result.deserialize(_apiClient.deserializeMultipartMap(_body));
+    _result.deserialize(_apiClient, _apiClient.deserializeMultipartMap(_body));
     return _result;
   }
 }