+++ lib/src/requests/apply_style_to_document_element_online_request.dart
@@ -120,7 +120,7 @@ class ApplyStyleToDocumentElementOnlineRequest implements RequestBase {
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
     var _result = ApplyStyleToDocumentElementOnlineResponse();
-    _result.deserialize(_apiClient.deserializeMultipartMap(_body));
+    _result.deserialize(_apiClient, _apiClient.deserializeMultipartMap(_body));
     return _result;
   }
 }