+++ lib/src/requests/delete_border_online_request.dart
@@ -114,7 +114,7 @@ class DeleteBorderOnlineRequest implements RequestBase {
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
     var _result = DeleteBorderOnlineResponse();
-    _result.deserialize(_apiClient.deserializeMultipartMap(_body));
+    _result.deserialize(_apiClient, _apiClient.deserializeMultipartMap(_body));
     return _result;
   }
 }