+++ lib/src/requests/delete_comments_online_request.dart
@@ -102,6 +102,5 @@ class DeleteCommentsOnlineRequest implements RequestBase {
 
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
-    return _body;
   }
 }