+++ lib/src/requests/delete_fields_online_request.dart
@@ -106,6 +106,5 @@ class DeleteFieldsOnlineRequest implements RequestBase {
 
   @override
   dynamic deserializeResponse(final ApiClient _apiClient, final ByteData _body) {
-    return _body;
   }
 }